cleanup version specific expectations
