+++ tests/PHPStan/Analyser/LegacyNodeScopeResolverTest.php
@@ -302,7 +302,7 @@ class LegacyNodeScopeResolverTest extends TypeInferenceTestCase
 				$testScope,
 				'matches',
 				TrinaryLogic::createYes(),
-				PHP_VERSION_ID <= 80000 ? 'array<string>' : 'array{0?: string}',
+				'array{0?: string}',
 			],
 			[
 				$testScope,
@@ -343,7 +343,7 @@ class LegacyNodeScopeResolverTest extends TypeInferenceTestCase
 				$testScope,
 				'matches2',
 				TrinaryLogic::createMaybe(),
-				PHP_VERSION_ID <= 80000 ? 'array<string>' : 'array{0?: string}',
+				'array{0?: string}',
 			],
 			[
 				$testScope,
@@ -355,7 +355,7 @@ class LegacyNodeScopeResolverTest extends TypeInferenceTestCase
 				$testScope,
 				'matches3',
 				TrinaryLogic::createYes(),
-				PHP_VERSION_ID <= 80000 ? 'array<string>' : 'array{0?: string}',
+				'array{0?: string}',
 			],
 			[
 				$testScope,
@@ -415,7 +415,7 @@ class LegacyNodeScopeResolverTest extends TypeInferenceTestCase
 				$testScope,
 				'ternaryMatches',
 				TrinaryLogic::createYes(),
-				PHP_VERSION_ID <= 80000 ? 'array<string>' : 'array{0?: string}',
+				'array{0?: string}',
 			],
 			[
 				$testScope,
