+++ tests/PHPStan/Analyser/nsrt/preg_match_shapes_php.php
